requirements.txt
google-cloud-bigquery>=3.14.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
requests>=2.31.0
